Moving call files:
- Call content file moved from voice folder to content folder
- Audio sources moved from content folder to audio-sources folder
